refactor(db) 7/13: extract db/npc.rs (closes #253)
Move 5 npc/companion/origin/combat-style populators, faction_from_fqn, their
CREATE TABLEs, and 5 tests into db/npc.rs. 240 tests + clippy -D warnings + fmt green.
